Set up automated wheel building and unittesting for macOS and (many)linux
Modified map_platform_specifics.py such that it can be run on its own (because we don't need makescript.py when using the cmake-based build).

The unittests.yml workflow builds and runs the pFUnit-based unittests and the pyTests. Currently, the test "test_critical_suite:test_critical_point_mix" fails on linux.

The cibuildwheels.yml workflow builds wheels for macOS x86_64 and arm64, and a manylinux2014 wheel, and runs the pyTests using the wheel for macOS arm64.

+++ addon/pycThermopack/makescript.py
@@ -35,26 +35,27 @@ def windows_make(diffs):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("mode", type=str, help="optim or debug")
-    parser.add_argument("-diffs", default="v3", type=str, help="Old (v2) or new (> v2) return mode for differentials", )
+    parser.add_argument("-diffs", default="v3", type=str, help="Old (v2) or new (> v2) return mode for differentials")
+    parser.add_argument("--install", action=argparse.BooleanOptionalAction, help="Install thermopack from thermopack/bin/dynamic?")
 
     args = parser.parse_args()
     mode = args.mode
 
-    # Glob to handle different OS suffixes
-    libthermo = list(
-        Path('../../bin/dynamic').glob(f'libthermopack_{mode}_gfortran*.*'))[0]
+    if args.install:
+        # Glob to handle different OS suffixes
+        libthermo = list(
+            Path('../../bin/dynamic').glob(f'libthermopack_{mode}_gfortran*.*'))[0]
+    
+        if not os.path.exists(libthermo):
+            print(f'{libthermo}does not exist. Have you compiled thermopack?')
+            sys.exit(1)
 
-    if not os.path.exists(libthermo):
-        print(f'{libthermo}does not exist. Have you compiled thermopack?')
-        sys.exit(1)
+        if os.path.exists("./thermopack/libthermopack"+libthermo.suffix):
+            os.remove("./thermopack/libthermopack"+libthermo.suffix)
 
-    if os.path.exists("./thermopack/libthermopack"+libthermo.suffix):
-        os.remove("./thermopack/libthermopack"+libthermo.suffix)
+        shutil.copy2(libthermo, "./thermopack/libthermopack"+libthermo.suffix)
 
-    shutil.copy2(libthermo, "./thermopack/libthermopack"+libthermo.suffix)
-
-    pf_specifics_path = os.path.join(os.path.dirname(
-        __file__), "thermopack", "platform_specifics.py")
+    pf_specifics_path = os.path.join(os.path.dirname(__file__), "thermopack", "platform_specifics.py")
     pf_specifics = map_platform_specifics.get_platform_specifics_by_trial_and_error()
 
     if args.diffs == 'v2':
@@ -83,7 +84,7 @@ if __name__ == "__main__":
     pf_specifics['diff_return_mode'] = args.diffs
     map_platform_specifics.write_platform_specifics_file(pf_specifics, pf_specifics_path)
 
-    map_platform_specifics.write_setup_file(f'v{version}')
-    map_platform_specifics.write_toml_file(version)
+    # map_platform_specifics.write_setup_file(f'v{version}')
+    # map_platform_specifics.write_toml_file(version)
 
     print(f'\033[92mSuccessfully configured ThermoPack {version}\033[0m')
